requirements.txt
openai>=1.0.0
openai-whisper
resemblyzer
pyaudio
pyttsx3
speechrecognition
numpy
pyyaml
python-dotenv
requests
beautifulsoup4
googletrans==4.0.0-rc1
yfinance
pywhatkit
google-auth
google-auth-oauthlib
google-api-python-client
spotipy
yt-dlp
pytesseract
opencv-python
qrcode
pillow
pdfplumber
speedtest-cli
psutil
paho-mqtt
paramiko
